No need to set foundValues more than only needed
It would affect minor performance while adding up the certain found values. For example, if a byte item with a picture could have many duplicate values like this:
A[0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,0,0,0,0,0,1,1,0,0,0,0,0,0,]

The change proposed will reduce the unnecessary and un-used variable changes in your code, boosting the efficiency when it comes to a very large input. 
And 1 minor typo correction in the comment.